Hopefully final refactor...
