+++ BOT/src/modules/default.config.json
