+++ BOT/src/modules/filescanner/actions.js
