+++ BOT/src/modules/filescanner/bin/scanner.js
